Implementation of model and note creation
It's a start to the model implementation using CoreData. So far, the
model contains a user that has notes. Notes contains tags and a folder.
Each entity is also obviously composed of basic type attributes (name,
id, token, dates, ...).

Secondly, it is now possible to create a note and add it to the
persistent store. Therefore, the changes are kept from one run to
another.

The menu has also been ordered in a better, more logical way.